chore(deps): bump everything, move to Astro 7 and Tailwind 4
Supersedes the open Dependabot PRs. The bumps are entangled, so they have
to land together: Astro 7 needs Vite 8, and `@astrojs/tailwind` peers at
`astro ^5`, so getting off Astro 5 forces the Tailwind 4 migration.

Dependencies:

- astro 5.18.1 -> 7.3.1, vite 7.3.5 -> 8.2.2 (a direct dep of astro 7)
- tailwindcss 3.4.19 -> 4.3.3, dropping `@astrojs/tailwind` for the
  `@tailwindcss/vite` plugin
- typescript 5.9.3 -> 6.0.3; 7 is out but `@astrojs/check` peers at
  `^5 || ^6`
- @types/node 22 -> 24 and `.node-version` 22 -> 24 to match; Node 24 is
  the current LTS line
- @astrojs/check, @astrojs/sitemap, @biomejs/biome and
  @tailwindcss/typography to latest

Tailwind 4:

- Ran `@tailwindcss/upgrade`, which deleted `tailwind.config.mjs` and
  moved the config into `src/assets/css/main.css`, renamed the utilities
  that changed name in v4 (`bg-gradient-to-*` -> `bg-linear-to-*`,
  `shadow-sm` -> `shadow-xs`, `flex-shrink-0` -> `shrink-0` and friends)
  and added the border-colour compatibility layer so v3's `gray-200`
  default is preserved.
- Dark mode is now a `@custom-variant dark` rule instead of
  `darkMode: "class"`.

Astro 7:

- Legacy content collections are gone: `src/content/config.ts` moves to
  `src/content.config.ts` and the collection takes a `glob()` loader. The
  glob keeps the old behaviour of skipping `_`-prefixed drafts.
- Content Layer renames: `post.slug` -> `post.id` and `post.render()` ->
  `render(post)`.

Also moves the pnpm settings out of `package.json` into
`pnpm-workspace.yaml`, which is where pnpm 10 reads them from, silencing
the warning on every command.

Verified `pnpm run check`, `pnpm run build` and `prek run --all-files`,
plus the rendered pages against a pre-upgrade capture: no horizontal
overflow at 375px, and the dark toggle and mobile menu still work.
