+++ pnpm-workspace.yaml
@@ -1,0 +1,4 @@
+onlyBuiltDependencies:
+  - '@biomejs/biome'
+  - esbuild
+  - sharp
