+++ src/assets/css/main.css
@@ -1,3 +1,27 @@
+@import "tailwindcss";
+
+@plugin "@tailwindcss/typography";
+
+@custom-variant dark (&:where(.dark, .dark *));
+
+/*
+  The default border color has changed to `currentcolor` in Tailwind CSS v4,
+  so we've added these compatibility styles to make sure everything still
+  looks the same as it did with Tailwind CSS v3.
+
+  If we ever want to remove these styles, we need to add an explicit border
+  color utility to any element that depends on these defaults.
+*/
+@layer base {
+  *,
+  ::after,
+  ::before,
+  ::backdrop,
+  ::file-selector-button {
+    border-color: var(--color-gray-200, currentcolor);
+  }
+}
+
 #sun {
   transform: translate3d(0, 0px, 0);
 }
@@ -15,6 +39,7 @@
 }
 
 html.dark #darkToggle:hover .horizon {
+  /* biome-ignore lint/complexity/noImportantStyles: needed to beat the .horizon border utility on hover */
   border-color: #718096 !important;
 }
