+++ src/components/header.astro
@@ -60,7 +60,7 @@ const menus = allMenus.filter((menu) => hasPosts || menu.url !== "/posts");
           class="absolute inset-0 top-0 right-0 block w-full h-full px-3 sm:hidden"
         >
           <div
-            class="relative w-full h-full bg-white border border-dashed border-neutral-300 dark:border-neutral-700 backdrop-blur-sm rounded-xl dark:bg-neutral-950"
+            class="relative w-full h-full bg-white border border-dashed border-neutral-300 dark:border-neutral-700 backdrop-blur-xs rounded-xl dark:bg-neutral-950"
           >
           </div>
         </div>
