+++ src/components/posts-loop.astro
@@ -12,7 +12,7 @@ const sortedPosts = allPosts.toSorted(
 const posts = sortedPosts.slice(0, count).map((post) => {
   return {
     ...(post.data || {}),
-    link: `/posts/${post.slug}`,
+    link: `/posts/${post.id}`,
   };
 });
 ---
@@ -35,7 +35,7 @@ const posts = sortedPosts.slice(0, count).map((post) => {
               {post.title}
             </a>
             <svg
-              class="group-hover:translate-x-0 flex-shrink-0 translate-y-0.5 -translate-x-1 w-2.5 h-2.5 stroke-current ml-1 transition-all ease-in-out duration-200  transform"
+              class="group-hover:translate-x-0 shrink-0 translate-y-0.5 -translate-x-1 w-2.5 h-2.5 stroke-current ml-1 transition-all ease-in-out duration-200  transform"
               viewBox="0 0 13 15"
               version="1.1"
               xmlns="http://www.w3.org/2000/svg"
