+++ src/components/square-lines.astro
@@ -7,7 +7,7 @@ import SquareLine from "./square-line.astro";
     class="absolute top-0 left-0 w-1/2 h-auto overflow-hidden bg-neutral-100 dark:bg-neutral-800"
   >
     <div
-      class="absolute inset-0 z-30 w-full h-full pointer-events-none bg-gradient-to-tl from-white dark:from-neutral-950 from-50% via-90% to-100% via-transparent to-transparent"
+      class="absolute inset-0 z-30 w-full h-full pointer-events-none bg-linear-to-tl from-white dark:from-neutral-950 from-50% via-90% to-100% via-transparent to-transparent"
     >
     </div>
     <div
@@ -26,7 +26,7 @@ import SquareLine from "./square-line.astro";
     class="absolute top-0 right-0 w-1/2 h-auto overflow-hidden bg-neutral-100 dark:bg-neutral-800"
   >
     <div
-      class="absolute inset-0 z-30 w-full h-full pointer-events-none bg-gradient-to-tr from-white dark:from-neutral-950 from-50% via-90% to-100% via-transparent to-transparent"
+      class="absolute inset-0 z-30 w-full h-full pointer-events-none bg-linear-to-tr from-white dark:from-neutral-950 from-50% via-90% to-100% via-transparent to-transparent"
     >
     </div>
     <div
