--- src/content/config.ts
@@ -1,14 +1,0 @@
-import { defineCollection, z } from "astro:content";
-
-const postsCollection = defineCollection({
-  type: "content",
-  schema: z.object({
-    title: z.string(),
-    description: z.string(),
-    date: z.date(),
-  }),
-});
-
-export const collections = {
-  posts: postsCollection,
-};
