+++ src/layouts/main.astro
@@ -4,6 +4,7 @@ import Footer from "../components/footer.astro";
 import Header from "../components/header.astro";
 import SquareLines from "../components/square-lines.astro";
 import "../assets/css/main.css";
+
 const { title } = Astro.props;
 
 // Only advertise the RSS feed when there is something in it
