+++ src/layouts/post.astro
@@ -12,11 +12,11 @@ const { post } = Astro.props;
       class="relative flex flex-col px-5 pt-6 border-t border-b-0 md:border-r md:border-l md:pt-20 lg:px-0 justify-stretch md:rounded-t-2xl border-neutral-200 dark:border-neutral-800"
     >
       <div
-        class="absolute top-0 left-0 hidden w-px h-full mt-1 -translate-x-px md:block bg-gradient-to-b from-transparent to-white dark:to-neutral-950"
+        class="absolute top-0 left-0 hidden w-px h-full mt-1 -translate-x-px md:block bg-linear-to-b from-transparent to-white dark:to-neutral-950"
       >
       </div>
       <div
-        class="absolute top-0 right-0 hidden w-px h-full mt-1 translate-x-px md:block bg-gradient-to-b from-transparent to-white dark:to-neutral-950"
+        class="absolute top-0 right-0 hidden w-px h-full mt-1 translate-x-px md:block bg-linear-to-b from-transparent to-white dark:to-neutral-950"
       >
       </div>
       <h1
