+++ src/pages/posts/rss.xml.ts
@@ -20,7 +20,7 @@ export async function GET(context: APIContext) {
         title: post.data.title,
         description: post.data.description,
         pubDate: post.data.date,
-        link: `/posts/${post.slug}`,
+        link: `/posts/${post.id}`,
       })),
   });
 }
