--- tailwind.config.mjs
@@ -1,9 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-  darkMode: "class",
-  content: ["./src/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [require("@tailwindcss/typography")],
-};
